Combine PUMA and CD helpers -> geo_helpers
No actual changes.
