+++ products/edde/aggregate/PUMS/pums_2000_demographics.py
@@ -6,7 +6,7 @@ from aggregate.aggregation_helpers import (
     order_aggregated_columns,
     get_category,
 )
-from utils.PUMA_helpers import dcp_pop_races
+from utils.geo_helpers import dcp_pop_races
 from internal_review.set_internal_review_file import set_internal_review_files
 from utils.dcp_population_excel_helpers import (
     race_suffix_mapper,
