+++ products/edde/aggregate/PUMS/pums_economics.py
@@ -3,7 +3,7 @@ from aggregate.aggregation_helpers import order_aggregated_columns
 from aggregate.load_aggregated import load_acs
 from internal_review.set_internal_review_file import set_internal_review_files
 from aggregate.PUMS.pums_2000_economics import pums_2000_economics
-from utils.PUMA_helpers import acs_years
+from utils.geo_helpers import acs_years
 from utils.dcp_population_excel_helpers import (
     race_suffix_mapper_global,
     count_suffix_mapper_global,
