+++ products/edde/aggregate/clean_aggregated.py
@@ -6,7 +6,7 @@ from utils.dcp_population_excel_helpers import (
     race_suffix_mapper_global,
     reorder_year_race,
 )
-from utils.PUMA_helpers import acs_years
+from utils.geo_helpers import acs_years
 
 
 def sort_columns(df: pd.DataFrame):
