+++ products/edde/aggregate/decennial_census/decennial_census_001020.py
@@ -2,7 +2,7 @@ from dcpy.utils.logging import logger
 from functools import cache
 import pandas as pd
 
-from utils.PUMA_helpers import clean_PUMAs
+from utils.geo_helpers import clean_PUMAs
 from internal_review.set_internal_review_file import set_internal_review_files
 
 # Map ACS year (or in general, input year for many functions) to decennial census year
