+++ products/edde/aggregate/housing_production/hpd_housing_ny_affordable_housing.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from utils.PUMA_helpers import borough_name_mapper, puma_from_coord
+from utils.geo_helpers import borough_name_mapper, puma_from_coord
 from ingest.ingestion_helpers import load_data
 
 cols = [
