+++ products/edde/aggregate/housing_security/DHS_shelter.py
@@ -3,8 +3,7 @@ but no CD. Something to watch out for when testing"""
 
 from aggregate.load_aggregated import initialize_dataframe_geo_index
 from internal_review.set_internal_review_file import set_internal_review_files
-from utils.CD_helpers import community_district_to_puma
-from utils.PUMA_helpers import borough_name_mapper
+from utils.geo_helpers import borough_name_mapper, community_district_to_puma
 from ingest import ingestion_helpers
 
 DATASET_NAME = "dhs_shelterd_indiv_by_comm_dist"
