+++ products/edde/aggregate/housing_security/eviction_cases_housing_court.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from internal_review.set_internal_review_file import set_internal_review_files
 
-from utils.CD_helpers import get_borough_num_mapper
+from utils.geo_helpers import borough_num_mapper
 
 
 def load_filings():
@@ -14,9 +14,7 @@ def load_filings():
     filings["citywide"] = "citywide"
 
     filings["Community District"] = filings["Community District"].astype(str)
-    filings["borough"] = (
-        filings["Community District"].str[0].map(get_borough_num_mapper())
-    )
+    filings["borough"] = filings["Community District"].str[0].map(borough_num_mapper)
     assert False, "Fix below before running"
     # filings = community_district_to_PUMA(
     #     df=filings, CD_col="Community District", CD_abbr_type="numeric_borough"
