+++ products/edde/aggregate/housing_security/households_rent_burden.py
@@ -7,7 +7,7 @@ from utils.dcp_population_excel_helpers import (
     race_suffix_mapper,
     count_suffix_mapper_global,
 )
-from utils.PUMA_helpers import acs_years
+from utils.geo_helpers import acs_years
 from internal_review.set_internal_review_file import set_internal_review_files
 from aggregate.load_aggregated import load_clean_housing_security_pop_data
 from aggregate.aggregation_helpers import get_geography_pop_data
