+++ products/edde/aggregate/housing_security/income_restricted_units.py
@@ -1,6 +1,6 @@
 import pandas as pd
 from internal_review.set_internal_review_file import set_internal_review_files
-from utils.PUMA_helpers import (
+from utils.geo_helpers import (
     clean_PUMAs,
     filter_for_recognized_pumas,
     puma_to_borough,
