+++ products/edde/aggregate/housing_security/nycha_tenants.py
@@ -2,7 +2,7 @@ from aggregate.decennial_census.decennial_census_001020 import decennial_census_
 from aggregate.aggregation_helpers import order_aggregated_columns
 import pandas as pd
 from internal_review.set_internal_review_file import set_internal_review_files
-from utils.PUMA_helpers import clean_PUMAs, puma_to_borough, acs_years
+from utils.geo_helpers import clean_PUMAs, puma_to_borough, acs_years
 from utils.dcp_population_excel_helpers import race_suffix_mapper_global
 
 SOURCE_DATA_FILE = (
