+++ products/edde/aggregate/housing_security/rent_median.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import numpy as np
-from utils.PUMA_helpers import acs_years
+from utils.geo_helpers import acs_years
 from internal_review.set_internal_review_file import set_internal_review_files
