+++ products/edde/aggregate/quality_of_life/access_to_jobs.py
@@ -1,6 +1,6 @@
 import pandas as pd
 from internal_review.set_internal_review_file import set_internal_review_files
-from utils.PUMA_helpers import clean_PUMAs, puma_to_borough
+from utils.geo_helpers import clean_PUMAs, puma_to_borough
 
 from ingest.ingestion_helpers import read_from_excel
