+++ products/edde/aggregate/quality_of_life/access_to_openspace.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from utils.PUMA_helpers import puma_to_borough, clean_PUMAs
+from utils.geo_helpers import puma_to_borough, clean_PUMAs
 from internal_review.set_internal_review_file import set_internal_review_files
 from ingest.ingestion_helpers import read_from_excel
