+++ products/edde/aggregate/quality_of_life/covid_death.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from utils.PUMA_helpers import puma_to_borough
+from utils.geo_helpers import puma_to_borough
 from internal_review.set_internal_review_file import set_internal_review_files
 import numpy as np
