+++ products/edde/aggregate/quality_of_life/health_mortality.py
@@ -4,7 +4,7 @@ import pandas as pd
 import numpy as np
 import re
 
-from utils.PUMA_helpers import clean_PUMAs
+from utils.geo_helpers import clean_PUMAs
 from internal_review.set_internal_review_file import set_internal_review_files
 from aggregate.clean_aggregated import order_PUMS_QOL_multiple_years
