+++ products/edde/aggregate/quality_of_life/heat_vulnerability.py
@@ -1,6 +1,6 @@
 import pandas as pd
 from internal_review.set_internal_review_file import set_internal_review_files
-from utils.PUMA_helpers import get_all_boroughs
+from utils.geo_helpers import get_all_boroughs
 
 
 def add_leading_zero_PUMA(df: pd.DataFrame) -> pd.DataFrame:
