+++ products/edde/aggregate/quality_of_life/traffic_fatalities.py
@@ -1,6 +1,6 @@
 import pandas as pd
 from internal_review.set_internal_review_file import set_internal_review_files
-from utils.PUMA_helpers import (
+from utils.geo_helpers import (
     get_all_NYC_PUMAs,
     clean_PUMAs,
     get_all_boroughs,
