+++ products/edde/external_review/collate_save_census.py
@@ -12,7 +12,7 @@ from aggregate.PUMS.pums_2000_demographics import pums_2000_demographics
 from aggregate.PUMS.pums_2000_economics import pums_2000_economics
 from aggregate.PUMS.pums_demographics import acs_pums_demographics
 from aggregate.PUMS.pums_economics import acs_pums_economics
-from utils.PUMA_helpers import acs_years
+from utils.geo_helpers import acs_years
 
 
 from aggregate.load_aggregated import initialize_dataframe_geo_index
