+++ products/edde/tests/general_indicator_tests/test_index_type.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 from tests.general_indicator_tests.general_indicator_test_helpers import get_by_geo
-from utils.PUMA_helpers import get_all_NYC_PUMAs, get_all_boroughs
+from utils.geo_helpers import get_all_NYC_PUMAs, get_all_boroughs
 
 all_PUMAs = get_all_NYC_PUMAs()
 all_boroughs = get_all_boroughs()
